requirements.txt
flask==2.0.0
flask-login==0.5.0
flask-migrate==3.0.0
flask-sqlalchemy==2.5.1
flask-wtf==0.14.3
python-dotenv==0.17.1
